+++ source/main/include/xbase/x_base.h
@@ -7,7 +7,7 @@
 
 namespace xbase
 {
-	extern void init(xcore::s32 number_of_threads=1, xcore::s64 temporary_allocator_size=8192);
+	extern void init(xcore::s32 number_of_threads=1, xcore::s32 temporary_allocator_size=8192);
 	extern void exit();
 };
